+++ src/circuit/state.jl
@@ -1,0 +1,7 @@
+# TODO: QuIDD
+
+type QuState{N}
+    s::AbstractVector
+end
+
+# TODO show(io::IO,state::QuState)
